Style cards
Sheesh, I don't even know anymore. After playing with gradients for
hours, I'm back to the plain, flat look, because all the gradient styles
I tried either looked dumb or weren't readable. Or both. But this is
kind of lame, so maybe I'll come back to it. I like it when things look
cool.